+++ INVENTARIO/src/main/java/INVENTARIO/dtos/Dispensa/Dispensa_Guardar.java
@@ -10,8 +10,8 @@ import java.io.Serializable;
 @Getter
 @Setter
 public class Dispensa_Guardar implements Serializable {
-    private Long prescripcionDetalleId;
-    private Short almacenId;
+    private Integer prescripcionDetalleId;
+    private Integer almacenId;
     private Integer loteId;
     private BigDecimal cantidad;
     private LocalDateTime fecha;
